+++ source/extensions/isaacsim.replicator.behavior/isaacsim/replicator/behavior/tests/__init__.py
@@ -8,3 +8,4 @@
 #
 
 from .test_behaviors_basic import *
+from .test_behaviors_sdg_scenario import *